refactor: add handleLine and handleClose
